% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCI_circle.r
\name{CCI_circle}
\alias{CCI_circle}
\title{Network viewing of cell-cell communication}
\usage{
CCI_circle(interaction_count, mycolor)
}
\arguments{
\item{interaction_count}{A data frame of the ligand-receptor interactions between pairwise cell types.}

\item{mycolor}{Colors used to represent different cell types.}
}
\value{
A network graph of the significant interactions.
}
\description{
Network viewing of cell-cell communication
}
