% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCI_network.r
\name{CCI_network}
\alias{CCI_network}
\title{Network viewing of cell-cell communication}
\usage{
CCI_network(
  interaction_count,
  mycolor,
  vertex.label.cex = 1.5,
  edge.label.cex = 1,
  title = "",
  edgeLabel = TRUE
)
}
\arguments{
\item{interaction_count}{A data frame of the ligand-receptor interactions between pairwise cell types.}

\item{mycolor}{Colors used to represent different cell types.}

\item{vertex.label.cex}{The label size of vertex. Default is 1.5.}

\item{edge.label.cex}{The size of label for arrows. Default is 1.}

\item{title}{The text for the title. Default is NULL.}

\item{edgeLabel}{Default is TRUE.}
}
\value{
A network graph of the significant interactions.
}
\description{
Network viewing of cell-cell communication
}
