% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_count.r
\name{heatmap_count}
\alias{heatmap_count}
\title{The number of ligand-receptor interactions between pairwise cell types}
\usage{
heatmap_count(
  interaction_count,
  log10 = FALSE,
  decimal = 1,
  text_size = 15,
  number_size = 5,
  title = NULL
)
}
\arguments{
\item{interaction_count}{A data frame of the ligand-receptor interactions between pairwise cell types.}

\item{log10}{Using log-count of interactions. Default is FALSE.}

\item{decimal}{Default is 1.}

\item{text_size}{Line/border size in mm; text size in pts. Default is 15.}

\item{number_size}{Font size for text annotations. Default is 5.}

\item{title}{The text for the title. Default is NULL.}
}
\value{
A figure of the significant interactions
}
\description{
This heatmap plot type includes two features count & log_count.
}
