% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_pathway.r
\name{LR_pathway}
\alias{LR_pathway}
\title{Construction of intracellular signaling pathway}
\usage{
LR_pathway(lr, objs, CellA, CellB, neighbor, geneSet, cor_value = 0.013, ...)
}
\arguments{
\item{lr}{A data frame of the ligand-receptor interactions.}

\item{objs}{object Seurat object.}

\item{CellA}{Ligands cell.}

\item{CellB}{Receptors cell.}

\item{neighbor}{Represents the second-order neighbors of ligand-receptor pair in the signaling network.Default is 2.}

\item{geneSet}{KEGG gene sets. (Ma et al., Nucleic Acids Research, 2021).}

\item{cor_value}{Spearman’s correlation between mean expression of ligand-receptor pair
and expression of each downstream gene. Default is 0.013.}

\item{...}{}
}
\value{
A data frame of the significantly enriched gene sets for each gene set category.
}
\description{
Construction of intracellular signaling pathway
}
