% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRTargetFisher.R
\name{LRTargetFisher}
\alias{LRTargetFisher}
\title{Functional category enrichment analysis}
\usage{
LRTargetFisher(Targets, GeneSets, TotalGene)
}
\arguments{
\item{Targets}{User input genes of interest.}

\item{GeneSets}{KEGG gene sets. (Ma et al., Nucleic Acids Research, 2021).}

\item{TotalGene}{Unique detected genes from a scRNA-seq dataset.}
}
\value{
A list. p value and overlapping genes.
}
\description{
Functional category enrichment analysis
}
